+++ code/src/app/Components/core/footer.tsx
